Changed UI of DropIn Checkout Buy Screen
